requirements.txt
Flask==0.12.1
Flask-Cors==3.0.2
Pillow==4.1.1
numpy==1.13.1
scikit-image==0.13.1
requests==2.18.1
